Created ComponentBase with code responsible for event handling and dispose
