+++ src/FoodPlannerBlazor/Components/NewCategoryComponent.razor
@@ -1,4 +1,6 @@
-﻿@page "/categories/new"
+﻿@inherits BaseComponent<NewCategoryComponentViewModel>
+
+@page "/categories/new"
 
 @if (showDetailsInformation)
 {
