+++ src/FoodPlannerBlazor/Components/NewCategoryComponent.razor.cs
@@ -1,16 +1,12 @@
-﻿using FoodPlannerBlazor.EditFormModels;
+﻿using FoodPlannerBlazor.Components.Common;
+using FoodPlannerBlazor.EditFormModels;
 using FoodPlannerBlazor.ViewModels;
-using Microsoft.AspNetCore.Components;
-using System.ComponentModel;
 using System.Threading.Tasks;
 
 namespace FoodPlannerBlazor.Components
 {
-    public partial class NewCategoryComponent : ComponentBase
+    public partial class NewCategoryComponent : BaseComponent<NewCategoryComponentViewModel>
     {
-        [Inject]
-        public NewCategoryComponentViewModel ViewModel { get; set; }
-
         private CreateCategoryFormModel formModel = new();
 
         private bool showDetailsInformation = false;
@@ -19,28 +15,6 @@ namespace FoodPlannerBlazor.Components
         {
             await ViewModel.AddCategoryAsync(formModel);
             showDetailsInformation = true;
-        }
-
-        public void Dispose() => ViewModel.PropertyChanged -= OnPropertyChangedHandler;
-
-        protected override async Task OnInitializedAsync()
-        {
-            ViewModel.PropertyChanged += async (sender, e) =>
-            {
-                await InvokeAsync(() =>
-                {
-                    StateHasChanged();
-                });
-            };
-            await base.OnInitializedAsync();
-        }
-
-        private async void OnPropertyChangedHandler(object sender, PropertyChangedEventArgs e)
-        {
-            await InvokeAsync(() =>
-            {
-                StateHasChanged();
-            });
         }
     }
 }
