+++ src/FoodPlannerBlazor/Components/NewPlannedMealComponent.razor
@@ -1,4 +1,6 @@
-﻿@page "/plannedMeals/new"
+﻿@inherits BaseComponent<NewPlannedMealComponentViewModel>
+
+@page "/plannedMeals/new"
 
 @if (showDetailsInformation)
 {
