+++ src/FoodPlannerBlazor/Components/PlannedMealsListComponent.razor
@@ -1,4 +1,6 @@
-﻿@page "/plannedMeals"
+﻿@inherits BaseComponent<PlannedMealsListComponentViewModel>
+
+@page "/plannedMeals"
 
 @if (IsStatic == false)
 {
