+++ src/FoodPlannerBlazor/_Imports.razor
@@ -9,6 +9,8 @@
 @using FoodPlannerBlazor
 @using FoodPlannerBlazor.Shared
 @using FoodPlannerBlazor.Components
+@using FoodPlannerBlazor.Components.Common
+@using FoodPlannerBlazor.ViewModels
 @using FoodPlannerBlazor.Infrastructure.Common
 @using FoodPlannerBlazor.Application.BusinessLogic.Category.Queries
 @using MediatR;
